Assignment complete.
I still dislike pig latin. Out of all the solutions here, that one
became the most hacky. Gross.

A lot of this was worked on in the earlier portion of the week of Java,
so I'm more than certain that a lot of these weren't the most optimal
way to do things.